+++ packages/app/hivecommand-backend/src/schema/devices/index.ts
@@ -160,8 +160,21 @@ export default (prisma: PrismaClient) => {
 								templatePacks: true,
 								alarms: {
 									include: {
-										conditions: true,
-										actions: true
+										edges: {
+											include: {
+												conditions: true
+											}
+										},
+										nodes: {
+											include: {
+												type: true,
+												targetedBy: {
+													include: {
+														source: true
+													}
+												}
+											}
+										}
 									}
 								},
 								components: {