+++ packages/app/hivecommand-native/src/views/controller/client.tsx
@@ -62,7 +62,6 @@ export const useLocalClient = (devices: any[]): CommandSurfaceClient => {
         //  }).reduce((prev, curr) => merge(prev, curr), {})
     }, [JSON.stringify(valueStore)])
 
-
     const onDataChanged = (data: any) => {
         console.log("data changed", data);
 
@@ -98,12 +97,10 @@ export const useLocalClient = (devices: any[]): CommandSurfaceClient => {
         socket.current.on('data-changed', onDataChanged)
 
 
-        return axios.get(`http://localhost:${8484}/${authState?.opcuaServer}/snapshot`).then((r) => r.data).then((data) => {
-            if (data.data) {
-                console.log("Initial state store", data.data.snapshot)
-                setValueStore(data.data.snapshot)
+        return axios.get(`http://localhost:${8484}/snapshot`).then((r) => r.data).then((data) => {
+            if (data) {
+                setValueStore(data.snapshot)
             }
-
         })
     }