+++ packages/communication/amqp-client/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.252](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.251...v1.4.15-alpha.252) (2023-06-18)
+
+**Note:** Version bump only for package @hive-command/amqp-client
+
+
+
+
+
+## [1.4.15-alpha.251](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.250...v1.4.15-alpha.251) (2023-06-18)
+
+**Note:** Version bump only for package @hive-command/amqp-client
+
+
+
+
+
+## [1.4.15-alpha.250](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.249...v1.4.15-alpha.250) (2023-06-18)
+
+**Note:** Version bump only for package @hive-command/amqp-client
+
+
+
+
+
+## [1.4.15-alpha.249](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.248...v1.4.15-alpha.249) (2023-06-08)
+
+**Note:** Version bump only for package @hive-command/amqp-client
+
+
+
+
+
+## [1.4.15-alpha.248](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.247...v1.4.15-alpha.248) (2023-06-08)
+
+**Note:** Version bump only for package @hive-command/amqp-client
+
+
+
+
+
 ## [1.4.15-alpha.247](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.246...v1.4.15-alpha.247) (2023-06-08)
 
 **Note:** Version bump only for package @hive-command/amqp-client