+++ packages/core/command-local-sidecar/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.256](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.255...v1.4.15-alpha.256) (2023-06-21)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
+## [1.4.15-alpha.255](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.254...v1.4.15-alpha.255) (2023-06-19)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
+## [1.4.15-alpha.254](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.253...v1.4.15-alpha.254) (2023-06-19)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
+## [1.4.15-alpha.253](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.252...v1.4.15-alpha.253) (2023-06-19)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
+## [1.4.15-alpha.252](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.251...v1.4.15-alpha.252) (2023-06-18)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
+## [1.4.15-alpha.251](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.250...v1.4.15-alpha.251) (2023-06-18)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
+## [1.4.15-alpha.250](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.249...v1.4.15-alpha.250) (2023-06-18)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
+## [1.4.15-alpha.249](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.248...v1.4.15-alpha.249) (2023-06-08)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
+## [1.4.15-alpha.248](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.247...v1.4.15-alpha.248) (2023-06-08)
+
+**Note:** Version bump only for package @hive-command/local-sidecar
+
+
+
+
+
 ## [1.4.15-alpha.247](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.246...v1.4.15-alpha.247) (2023-06-08)
 
 **Note:** Version bump only for package @hive-command/local-sidecar