+++ packages/drivers/command-opcua/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.256](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.255...v1.4.15-alpha.256) (2023-06-21)
+
+**Note:** Version bump only for package @hive-command/drivers-opcua
+
+
+
+
+
+## [1.4.15-alpha.255](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.254...v1.4.15-alpha.255) (2023-06-19)
+
+**Note:** Version bump only for package @hive-command/drivers-opcua
+
+
+
+
+
+## [1.4.15-alpha.254](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.253...v1.4.15-alpha.254) (2023-06-19)
+
+**Note:** Version bump only for package @hive-command/drivers-opcua
+
+
+
+
+
+## [1.4.15-alpha.253](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.252...v1.4.15-alpha.253) (2023-06-19)
+
+**Note:** Version bump only for package @hive-command/drivers-opcua
+
+
+
+
+
 ## [1.4.15-alpha.234](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.233...v1.4.15-alpha.234) (2023-06-02)
 
 **Note:** Version bump only for package @hive-command/drivers-opcua